sub2/backend/requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
branca==0.3.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
folium==0.10.1
gast==0.2.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.8
isort==4.3.21
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==0.25.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.13.0
sqlparse==0.3.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.7
Werkzeug==1.0.0
wrapt==1.11.2
